Rename resource to handler
