+++ service/controller/handler/test/create.go
@@ -1,0 +1,9 @@
+package test
+
+import (
+	"context"
+)
+
+func (r *Handler) EnsureCreated(ctx context.Context, obj interface{}) error {
+	return nil
+}
