+++ service/controller/handler/test/delete.go
@@ -1,0 +1,9 @@
+package test
+
+import (
+	"context"
+)
+
+func (r *Handler) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}
