--- service/controller/resource/test/create.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"context"
-)
-
-func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	return nil
-}
