--- service/controller/resource/test/delete.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"context"
-)
-
-func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	return nil
-}
